move language_switch hyperscript to jsx
