+++ euth/contrib/static/js/app.js
@@ -1,6 +1,6 @@
 var ReactComments = require('../../../comments/static/comments/react_comments.js')
 var ReactRatings = require('../../../ratings/static/ratings/react_ratings.js')
-var ReactLanguageSwitch = require('../../../dashboard/static/language_switch/react_language_switch.js')
+var ReactLanguageSwitch = require('../../../dashboard/static/language_switch/react_language_switch.jsx')
 
 module.exports = {
   'renderComment': ReactComments.renderComment,
